requirements.txt
autoenv==1.0.0
certifi==2018.8.24
chardet==3.0.4
click==6.7
Click>=6.0
dlib>=19.3.0
face_recognition_models
Flask-JWT==0.3.2
flask-oidc==1.4.0
Flask==1.0.2
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy
oauth2client==4.1.3
Pillow
pyasn1-modules==0.2.2
pyasn1==0.4.4
PyJWT==1.4.2
python-dotenv==0.9.1
requests==2.19.1
rsa==4.0
scipy>=0.17.0
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
